+++ src/main/java/name/valery1707/validator/Checker.java
@@ -1,5 +1,7 @@
 package name.valery1707.validator;
 
+import java.util.regex.Pattern;
+
 public final class Checker {
 	Checker() {
 		throw new IllegalStateException("Instance must not be created");
@@ -16,6 +18,14 @@ public final class Checker {
 	public static int checkRange(int value, int min, int max, String name) {
 		if (value < min || value > max) {
 			throw new IllegalArgumentException(String.format("%s must be between %d and %d", name, min, max));
+		} else {
+			return value;
+		}
+	}
+
+	public static String checkPattern(String value, Pattern pattern, String name) {
+		if (!pattern.matcher(value).matches()) {
+			throw new IllegalArgumentException(String.format("%s must match with pattern %s", name, pattern.toString()));
 		} else {
 			return value;
 		}